+++ CHANGELOG.md
@@ -68,6 +68,7 @@ Release: dd.mm.yyyy
 #### Window
 
 - Change: Function `focusedWindow()` is now simply `focused()` ([#108](https://github.com/kasper/phoenix/issues/108)).
+- Change: Function `windows()` is now `all()` ([#108](https://github.com/kasper/phoenix/issues/108)).
 
 #### Command